v5.0.5: 🛠️ 整理了一些代码

+++ PlainCEETimer.Natives/MemoryCleaner/MemoryCleaner.h
@@ -3,4 +3,4 @@
 #include <windows.h>
 #include <psapi.h>
 
-cexport(void) CleanMemory(int threshold);
+cexport(void) CleanMemory(SIZE_T threshold);
